backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0,<4.0
django-cors-headers>=4.3.1,<5.0
psycopg2-binary>=2.9.9,<3.0
pytest>=7.4.0,<8.0
pytest-django>=4.5.2,<5.0
gunicorn>=21.2.0,<22.0
